+++ operator/pkg/controllers/cluster_cleanup.go
@@ -16,14 +16,11 @@ import (
 	"github.com/ais-operator/pkg/resources/statsd"
 	"github.com/ais-operator/pkg/resources/target"
 	batchv1 "k8s.io/api/batch/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 )
 
 func (r *AIStoreReconciler) cleanup(ctx context.Context, ais *aisv1.AIStore) (updated bool, err error) {
-	nodeNames, err := r.k8sClient.ListNodesRunningAIS(ctx, ais)
-	if err != nil {
-		r.log.Error(err, "Failed to list nodes running AIS")
-	}
 	updated, err = cmn.AnyFunc(
 		func() (bool, error) { return r.cleanupTarget(ctx, ais) },
 		func() (bool, error) { return r.cleanupProxy(ctx, ais) },
@@ -33,28 +30,26 @@ func (r *AIStoreReconciler) cleanup(ctx context.Context, ais *aisv1.AIStore) (up
 		func() (bool, error) { return r.cleanupPVC(ctx, ais) },
 		func() (bool, error) { return r.cleanupTLS(ctx, ais) },
 	)
-	if updated && ais.ShouldCleanupMetadata() {
-		err = r.createCleanupJobs(ctx, ais, nodeNames)
-		if err != nil {
-			return
-		}
-		err = r.updateStatusWithState(ctx, ais, aisv1.HostCleanup)
-		if err != nil {
-			return
-		}
+	if err != nil {
+		return
+	}
+	// If not using storage class, trigger jobs to clean up host mounts
+	// Note this is not part of the above list to avoid host cleanup jobs blocking any K8s resource cleanup
+	if ais.ShouldCleanupMetadata() && ais.Spec.StateStorageClass == nil {
+		err = r.cleanupHostStateMounts(ctx, ais)
 	}
 	return
 }
 
 func (r *AIStoreReconciler) createCleanupJobs(ctx context.Context, ais *aisv1.AIStore, nodes []string) error {
-	if ais.Spec.StateStorageClass != nil {
-		return nil
-	}
 	logger := logf.FromContext(ctx)
 	logger.Info("Creating manual cleanup jobs", "nodes", nodes)
 	for _, nodeName := range nodes {
 		jobDef := cmn.NewCleanupJob(ais, nodeName)
 		if err := r.k8sClient.Create(ctx, jobDef); err != nil {
+			if apierrors.IsAlreadyExists(err) {
+				continue
+			}
 			logger.Error(err, "Failed to create cleanup job", "name", jobDef.Name, "node", nodeName)
 			return err
 		}
@@ -111,17 +106,27 @@ func (r *AIStoreReconciler) deleteFinishedJobs(ctx context.Context, jobs *batchv
 	return remaining, nil
 }
 
+func (r *AIStoreReconciler) cleanupHostStateMounts(ctx context.Context, ais *aisv1.AIStore) error {
+	nodeNames, err := r.k8sClient.ListNodesMatchingAISSelectors(ctx, ais)
+	if err != nil {
+		return err
+	}
+	err = r.createCleanupJobs(ctx, ais, nodeNames)
+	if err != nil {
+		return err
+	}
+	return r.updateStatusWithState(ctx, ais, aisv1.HostCleanup)
+}
+
 func (r *AIStoreReconciler) cleanupPVC(ctx context.Context, ais *aisv1.AIStore) (bool, error) {
+	// Note: Data PVCs must be deleted in-sync with state PVCs
+	// Otherwise we risk new PVs being created on nodes with conflicting pre-existing state,
+	// including cluster maps with ordinal-based addresses, e.g. ais-target-0.ais.svc.cluster.local
 	if !ais.ShouldCleanupMetadata() {
 		return false, nil
 	}
-	if ais.Spec.CleanupData != nil && *ais.Spec.CleanupData {
-		return r.deleteAllPVCs(ctx, ais)
-	}
-	if ais.Spec.StateStorageClass != nil {
-		return r.deleteStatePVCs(ctx, ais)
-	}
-	return false, nil
+	// Includes all PVCs (both data and state)
+	return r.deleteAllPVCs(ctx, ais)
 }
 
 func (r *AIStoreReconciler) deleteAllPVCs(ctx context.Context, ais *aisv1.AIStore) (bool, error) {
@@ -132,17 +137,6 @@ func (r *AIStoreReconciler) deleteAllPVCs(ctx context.Context, ais *aisv1.AIStor
 	}
 	r.log.Info("Cleaning up all proxy PVCs")
 	return r.k8sClient.DeletePVCs(ctx, ais.Namespace, proxy.RequiredPodLabels(ais), nil)
-}
-
-// Cleans up only dynamically created volumes by adding a filter by the defined state storage class
-func (r *AIStoreReconciler) deleteStatePVCs(ctx context.Context, ais *aisv1.AIStore) (bool, error) {
-	r.log.Info("Cleaning up dynamic target PVCs")
-	updated, err := r.k8sClient.DeletePVCs(ctx, ais.Namespace, target.RequiredPodLabels(ais), ais.Spec.StateStorageClass)
-	if err != nil {
-		return updated, err
-	}
-	r.log.Info("Cleaning up dynamic proxy PVCs")
-	return r.k8sClient.DeletePVCs(ctx, ais.Namespace, proxy.RequiredPodLabels(ais), ais.Spec.StateStorageClass)
 }
 
 func (r *AIStoreReconciler) cleanupRBAC(ctx context.Context, ais *aisv1.AIStore) (anyUpdated bool, err error) {